deep-research/novel-generator/pyproject.toml
[project]
name = "teen-novel-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipython>=9.4.0",
    "openai>=1.97.1",
    "openai-agents>=0.2.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sendgrid>=6.12.4",
]
